requirements.txt
amqp==2.6.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.9
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.5
click==7.1.2
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
kombu==4.6.10
Markdown==3.2.2
mccabe==0.6.1
pathspec==0.8.0
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
regex==2020.6.8
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
vine==1.3.0
whitenoise==5.1.0
